pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.5",
    "opencv-python-headless>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "plotly>=6.0.1",
    "pytesseract>=0.3.13",
    "streamlit>=1.45.0",
]
